telegram-bot/requirements.txt
aiogram>=3.23.0
aiohttp>=3.9.0
python-dotenv>=1.0.0


